+++ migrations/123_zones__number_grade__add.sql
@@ -1,1 +1,0 @@
-ALTER TABLE zones ADD COLUMN IF NOT EXISTS number_grade INTEGER;