+++ pages/studentAssessmentInstanceExam/studentAssessmentInstanceExam.sql
@@ -10,8 +10,8 @@ SELECT
     qo.question_number,
     z.max_points AS zone_max_points,
     (z.max_points IS NOT NULL) AS zone_has_max_points,
-    z.number_grade AS zone_number_grade,
-    (z.number_grade IS NOT NULL) AS zone_has_number_grade
+    z.best_questions AS zone_best_questions,
+    (z.best_questions IS NOT NULL) AS zone_has_best_questions
 FROM
     instance_questions AS iq
     JOIN assessment_instances AS ai ON (ai.id = iq.assessment_instance_id)