+++ sync/fromDisk/assessments.js
@@ -215,7 +215,7 @@ module.exports = {
                 title: dbZone.title,
                 number_choose: dbZone.numberChoose,
                 max_points: dbZone.maxPoints,
-                number_grade: dbZone.numberGrade,
+                best_questions: dbZone.bestQuestions,
             };
             sqldb.queryOneRow(sql.insert_zone, params, function(err, result) {
                 if (ERR(err, callback)) return;